requirements.txt
Pillow
watchdog
pyyaml

